Add "random communication benchmark" tool
git-svn-id: https://svn.apache.org/repos/asf/incubator/hama/trunk@1072853 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hama/bsp/RunningJob.java
@@ -60,7 +60,7 @@ public interface RunningJob {
    * @return the progress of the job's tasks.
    * @throws IOException
    */
-  public float progress() throws IOException;
+  public long progress() throws IOException;
 
   /**
    * Check if the job is finished or not. This is a non-blocking call.
